Remove the "resize-pane" command in new_pane
After restoring the pane contents and command, calling "resize-pane" on the
window has the effect of clearing the contents of all but the last pane, which
is why my page contents wasn't restoring correctly.

It's not clear what the intent of this command was, it was commented "minimize
window so more panes can fit" but I don't see why it should matter whether
"more panes fit" since we're adding them with split-window.  In any case,
things seem to work fine (better, in fact) without it.